修改

+++ ancun-netty-up2yun/src/main/java/com/ancun/up2yun/cfg/NettyServerConfig.java
@@ -33,7 +33,7 @@ public class NettyServerConfig {
     private int tcpPort;
 
     @Value("${upload.file.size}")
-    private long maxContentSize;
+    private int maxContentSize;
 
     @Value("${so.keepalive}")
     private boolean keepAlive;
